Factor out UI logic testing utils
